+++ src/app/Contracts/CustomFilter.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace LaravelEnso\Tables\app\Contracts;
+
+use Illuminate\Database\Eloquent\Builder;
+
+interface CustomFilter
+{
+    public function filter(Builder $query);
+}
