+++ tests/units/Services/Table/Filters/SearchTest.php
@@ -1,16 +1,13 @@
 <?php
 
-namespace LaravelEnso\Tables\Tests\units\Services\Table;
+namespace LaravelEnso\Tables\Tests\units\Services\Table\Filters;
 
 use Faker\Factory;
 use Tests\TestCase;
-use Illuminate\Support\Facades\Schema;
-use Illuminate\Database\Eloquent\Model;
-use LaravelEnso\Helpers\app\Classes\Obj;
-use LaravelEnso\Tables\app\Services\Table\Filters;
 use LaravelEnso\Tables\app\Services\Table\Request;
+use LaravelEnso\Tables\app\Services\Table\Filters\Search;
 
-class FilterTest extends TestCase
+class SearchTest extends TestCase
 {
     private $testModel;
     private $faker;
@@ -25,13 +22,23 @@ class FilterTest extends TestCase
 
         $this->faker = Factory::create();
 
-        $this->createTestModelTable();
+        TestModel::createTable();
 
-        $this->createRelationModelTable();
+        RelationalModel::createTable();
 
         $this->testModel = $this->createTestModel();
-        $this->query = FilterTestModel::select('*');
-        $this->params = ['meta' => []];
+
+        $this->query = TestModel::select('*');
+
+        $this->params = [
+            'meta' => [],
+            'columns' => [
+                'name' => [
+                    'data' => 'name',
+                    'meta' => ['searchable' => true]
+                ]
+            ]
+        ];
 
         $this->createRelationModel();
     }
@@ -43,7 +50,7 @@ class FilterTest extends TestCase
 
         $response = $this->requestResponse();
 
-        $this->assertCount(FilterTestModel::count(), $response);
+        $this->assertCount(TestModel::count(), $response);
 
         $this->assertTrue(
             $response->pluck('name')
@@ -54,11 +61,6 @@ class FilterTest extends TestCase
     /** @test */
     public function can_use_search()
     {
-        $this->params['columns']['name'] = [
-            'data' => 'name',
-            'meta' => ['searchable' => true]
-        ];
-
         $this->params['meta'] = [
             'search' => $this->testModel->name,
             'comparisonOperator' => 'LIKE',
@@ -84,11 +86,6 @@ class FilterTest extends TestCase
     /** @test */
     public function can_use_full_search()
     {
-        $this->params['columns']['name'] = [
-            'data' => 'name',
-            'meta' => ['searchable' => true]
-        ];
-
         $this->params['meta'] = [
             'search' => substr($this->testModel->name, 1, strlen($this->testModel->name) - 2),
             'comparisonOperator' => 'LIKE',
@@ -114,11 +111,6 @@ class FilterTest extends TestCase
     /** @test */
     public function can_use_starts_with_search()
     {
-        $this->params['columns']['name'] = [
-            'data' => 'name',
-            'meta' => ['searchable' => true]
-        ];
-
         $this->params['meta'] = [
             'search' => collect(explode(' ', $this->testModel->name))->first(),
             'comparisonOperator' => 'LIKE',
@@ -138,11 +130,6 @@ class FilterTest extends TestCase
     /** @test */
     public function can_use_ends_with_search()
     {
-        $this->params['columns']['name'] = [
-            'data' => 'name',
-            'meta' => ['searchable' => true]
-        ];
-
         $this->params['meta'] = [
             'search' => collect(explode(' ', $this->testModel->name))->last(),
             'comparisonOperator' => 'LIKE',
@@ -168,17 +155,10 @@ class FilterTest extends TestCase
     /** @test */
     public function can_use_multi_argument_search()
     {
-        $this->params['columns'] = [
-            'name' => [
-                'data' => 'name',
-                'searchable' => true,
-                'meta' => ['searchable' => true]
-            ],
-            'appellative' => [
-                'data' => 'appellative',
-                'searchable' => true,
-                'meta' => ['searchable' => true]
-            ],
+        $this->params['columns']['appellative'] = [
+            'data' => 'appellative',
+            'searchable' => true,
+            'meta' => ['searchable' => true]
         ];
 
         $this->params['meta'] = [
@@ -222,88 +202,9 @@ class FilterTest extends TestCase
         );
     }
 
-    /** @test */
-    public function can_use_interval()
-    {
-        $this->params['intervals']['filter_test_models']['id'] = [
-            'min' => $this->testModel->id - 1,
-            'max' => $this->testModel->id + 1,
-        ];
-
-        $response = $this->requestResponse();
-
-        $this->assertCount(1, $response);
-
-        $this->assertEquals(
-            $this->testModel->name,
-            $response->first()->name
-        );
-
-        $this->params['intervals']['filter_test_models']['id'] = [
-            'min' => $this->testModel->id - 1,
-            'max' => $this->testModel->id - 2,
-        ];
-
-        $response = $this->requestResponse();
-
-        $this->assertCount(0, $response);
-    }
-
-    /** @test */
-    public function can_use_date_interval()
-    {
-        $this->params['intervals']['filter_test_models']['created_at'] = [
-            'dbDateFormat' => 'Y-m-d',
-            'dateFormat' => 'Y-m-d',
-            'min' => $this->testModel->created_at->subDays(1)->format('Y-m-d'),
-            'max' => $this->testModel->created_at->addDays(1)->format('Y-m-d'),
-        ];
-
-        $response = $this->requestResponse();
-
-        $this->assertCount(1, $response);
-
-        $this->assertEquals(
-            $this->testModel->name,
-            $response->first()->name
-        );
-
-        $this->params['intervals']['filter_test_models']['created_at'] = [
-            'dbDateFormat' => 'Y-m-d',
-            'dateFormat' => 'Y-m-d',
-            'min' => $this->testModel->created_at->subDays(2)->format('Y-m-d'),
-            'max' => $this->testModel->created_at->subDays(1)->format('Y-m-d'),
-        ];
-
-        $response = $this->requestResponse();
-
-        $this->assertCount(0, $response);
-    }
-
-    /** @test */
-    public function can_use_filters()
-    {
-        $this->params['filters']['filter_test_models'] = ['name' => $this->testModel->name];
-
-        $response = $this->requestResponse();
-
-        $this->assertCount(1, $response);
-
-        $this->assertEquals(
-            $response->first()->name,
-            $this->testModel->name
-        );
-
-        $this->params['filters']['filter_test_models'] = ['name' => $this->testModel->name.'-'];
-
-        $response = $this->requestResponse();
-
-        $this->assertCount(0, $response);
-    }
-
     private function requestResponse()
     {
-        (new Filters(
+        (new Search(
             new Request($this->params),
             $this->query
         ))->handle();
@@ -313,7 +214,7 @@ class FilterTest extends TestCase
 
     private function createTestModel()
     {
-        return FilterTestModel::create([
+        return TestModel::create([
             'appellative' => $this->faker->firstName,
             'name' => $this->faker->name,
         ]);
@@ -321,45 +222,9 @@ class FilterTest extends TestCase
 
     private function createRelationModel()
     {
-        return FilterRelationModel::create([
+        return RelationalModel::create([
             'name' => $this->faker->word,
             'parent_id' => $this->testModel->id,
         ]);
     }
-
-    private function createTestModelTable()
-    {
-        Schema::create('filter_test_models', function ($table) {
-            $table->increments('id');
-            $table->string('appellative')->nullable();
-            $table->string('name')->nullable();
-            $table->timestamps();
-        });
-    }
-
-    private function createRelationModelTable()
-    {
-        Schema::create('filter_relation_models', function ($table) {
-            $table->increments('id');
-            $table->string('name')->nullable();
-            $table->integer('parent_id')->nullable();
-            $table->foreign('parent_id')->references('id')->on('filter_test_models');
-            $table->timestamps();
-        });
-    }
-}
-
-class FilterTestModel extends Model
-{
-    protected $fillable = ['name', 'appellative', 'created_at'];
-
-    public function relation()
-    {
-        return $this->hasOne(FilterRelationModel::class, 'parent_id');
-    }
-}
-
-class FilterRelationModel extends Model
-{
-    protected $fillable = ['name', 'parent_id'];
 }
